requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
cycler==0.10.0
deepsig==1.0.1
fasttext==0.9.2
joblib==1.0.1; python_version >= '3.6'
kiwisolver==1.3.1; python_version >= '3.6'
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
pillow==8.2.0; python_version >= '3.6'
python-dotenv==0.17.1
pytz==2021.1
scikit-learn==0.24.2; python_version >= '3.6'
scipy==1.5.4
seaborn==0.11.1
sklearn==0.0
threadpoolctl==2.1.0; python_version >= '3.5'
torch==1.8.1
tqdm==4.59.0
typing-extensions==3.10.0.0
